backend/requirements.txt
# ============================================================
# Dépendances VoiceClonIA — backend Python
# Environnement cible : RTX 4090 / CUDA 12.8 / PyTorch 2.8
#
# Ordre d'installation recommandé :
#   1. pip install torch torchaudio --index-url https://download.pytorch.org/whl/cu128
#   2. pip install chatterbox-tts
#   3. pip install -r backend/requirements.txt
# ============================================================

# --- Serveur web ---
fastapi==0.115.6
uvicorn[standard]==0.32.1

# --- Base de données ---
sqlalchemy==2.0.36

# --- Variables d'environnement ---
python-dotenv==1.0.1

# --- Validation des données ---
pydantic==2.10.3

# --- Traitement audio ---
librosa>=0.10.2
soundfile==0.12.1
numpy==1.26.4

# --- Upload de fichiers multipart ---
python-multipart==0.0.12

# --- Client HTTP (Voxtral/vLLM Omni) ---
httpx>=0.27.0

# --- PyTorch — NE PAS installer via ce fichier ---
# PyTorch 2.8 + CUDA 12.8 doit être installé AVANT les TTS engines :
#   pip install torch torchaudio --index-url https://download.pytorch.org/whl/cu128

# --- Moteur TTS principal : Chatterbox (Resemble AI) ---
# pip install chatterbox-tts
# Note : chatterbox-tts détecte automatiquement CUDA — RTX 4090 = inférence ~temps réel

# --- Moteur TTS alternatif : F5-TTS ---
# pip install f5-tts
